addressed the mentor requested changes
Disabled Nav buttons at boundries instead of hiding them , Go to Page redirects to nearest valid value , addressed duplication and state leakage
