+++ package.json
@@ -4,7 +4,7 @@
     "description": "GSoC 2026 Organisation Finder",
     "private": true,
     "scripts": {
-        "lint:js": "eslint src/js/app.js src/js/org.js src/js/tailwind-config.js agent/scripts/validate-ideas-urls.js agent/scripts/extract-mentors.js agent/scripts/validate-mentors.js api/github.js src/js/githubAnalyzer.js src/js/skillExtractor.js src/js/recommender.js src/js/recommendation-ui.js src/js/footer.js",
+        "lint:js": "eslint src/js/app.js src/js/pagination.js src/js/org.js src/js/tailwind-config.js agent/scripts/validate-ideas-urls.js agent/scripts/extract-mentors.js agent/scripts/validate-mentors.js api/github.js src/js/githubAnalyzer.js src/js/skillExtractor.js src/js/recommender.js src/js/recommendation-ui.js src/js/footer.js",
         "lint:css": "stylelint src/styles.css src/landing.css",
         "lint:html": "htmlhint --config .htmlhintrc index.html privacy.html landing.html",
         "lint": "npm run lint:js && npm run lint:css && npm run lint:html",
