+++ tests/browser.test.js
@@ -3,6 +3,19 @@ const assert = require('node:assert');
 
 require('./helpers/setup-globals.js');
 
+// Snapshot the globals this suite overrides so we can put them back afterwards.
+// Without this, the custom window/document/location/fetch below leak into any
+// test file that runs later in the same process, causing order-dependent fails.
+const _savedGlobals = {
+  window: globalThis.window,
+  document: globalThis.document,
+  location: globalThis.location,
+  fetch: globalThis.fetch
+};
+test.after(() => {
+  Object.assign(globalThis, _savedGlobals);
+});
+
 // Custom window/location for browser DOM tests (overrides the helper's stub)
 globalThis.window = {
   addEventListener: () => {},
